#16 - Fix two key bugs
Was changing data on rising edge (incorrect), not falling edge (correct)

Previously, in spin wait loop, was terminating as soon as last word was sampled - but this is done before the last edge of the TCK was changed!  Now swapped order - change TCK first (rising edge sample), and then sample.